backend/requirements.txt
Flask==3.0.2
moviepy==1.0.3
werkzeug>=2.0.0
numpy>=1.21.0
imageio>=2.9.0
imageio-ffmpeg>=0.4.5
tqdm>=4.62.0
decorator>=4.0.2,<5.0.0
requests==2.31.0
openai==0.28.1
ffmpeg-python>=0.2.0
nltk==3.8.1
sentence-transformers==2.2.2
transformers>=4.30.0
torch>=2.0.0
sentencepiece>=0.1.99
protobuf>=3.20.0
flask-cors==4.0.0
python-dotenv==1.0.1
huggingface-hub==0.21.4

# Core dependencies
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.27
faiss-cpu==1.7.4
altair>=4.0.0
tiktoken>=0.4.0
InstructorEmbedding>=1.0.1

whisper==1.1.1